Replace exclusive-edit with transcluded comment directive
This commit removes the somewhat superfluous exlusiveEditController
in favor of event-based passing of quoted text to the correct comment
field.

The comment directive now decides for itself whether it wants to be
placed above or below the activities and uses transclusion for that
purpose.

Editing comments is now no longer exlusive, but that shouldn't be
necessary anyway.