Parallel sending of comment and fields in WP split view
Parallel submitting of comments alongside WP edits wasn't possible
natively due to the increased lock_version in work packages, caused by
`touch`ing and explicitly saving the work package in the process of
adding a journal.

This commit works around this issue by updating the timestamp columns
directly, bypassing optimistic locking in the work package table.

This allows comments to be submitted transparently with other fields to
the user, while actually, we send it in two subsequent steps.